Don't change player match count for bye matches
Before changing a player's match count, check to see if we're using a
bye match first. Because functions that change matches are spread across
different modules, the code is still a bit messy. We may need a more
robust strategy to update match counts to guarantee consistency.

This also changes the type for match counts from int to a natural-number
int that can't go below zero.

Fixes #83